requirements.txt
gdown
matplotlib==3.5.1
numpy
opencv_python==4.5.5.64
opencv-contrib-python==4.6.0.66
tqdm==4.64.0
pillow
torchvision
datasets
timm==0.5.4
transformers
scikit-learn
matplotlib
evaluate
accelerate
wandb
jupyter
albumentations
einops